% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kaggle.R
\name{calculate_relative_submissions}
\alias{calculate_relative_submissions}
\title{Transform submissions to be relative to the first place team.}
\usage{
calculate_relative_submissions(leaderboards)
}
\description{
Requires first place team to be labeled (FirstPlaceTeam columnn).
See \code{\link{label_place_groups}}.
}

