% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kaggle.R
\name{calculate_submission_intervals}
\alias{calculate_submission_intervals}
\title{Compute time interval between first and last submissions.}
\usage{
calculate_submission_intervals(submissions)
}
\description{
Compute time interval between first and last submissions.
}

