% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kaggle.R
\name{connect_kaggle}
\alias{connect_kaggle}
\title{Connect to the Kaggle SQLite database.}
\usage{
connect_kaggle(sqlite_location)
}
\arguments{
\item{sqlite_location}{Location of Kaggle sqlite database. Download this file
from \url{https://www.kaggle.com/kaggle/meta-kaggle/downloads/database.sqlite.zip},
unzip it, and put it wherever you'd like. By default, the file is expected
to be found in the "evoteams" package at "inst/extdata/kaggle.sqlite".}
}
\description{
Connect to the Kaggle SQLite database.
}
\examples{
kaggle_db <- connect_kaggle("path/to/kaggle.sqlite")
kaggle_db <- connect_kaggle()  # assumes db included in evoteams pkg

}

