% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kaggle.R
\name{divide_into_team_types}
\alias{divide_into_team_types}
\title{Assign team types by median splits on time and submissions.}
\usage{
divide_into_team_types(leaderboards)
}
\description{
See also \code{\link{recode_team_type}}.
}

