% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kaggle.R
\name{get_competition_intervals}
\alias{get_competition_intervals}
\title{Get competition start and end times.}
\usage{
get_competition_intervals(kaggle_db)
}
\description{
Get competition start and end times.
}

