% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kaggle.R
\name{get_place_mod_preds}
\alias{get_place_mod_preds}
\title{Get model predictions for a model predicting Place from TotalSubmissions.}
\usage{
get_place_mod_preds(mod, predict_fn, x_preds)
}
\description{
Get model predictions for a model predicting Place from TotalSubmissions.
}

