% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kaggle.R
\name{get_submissions}
\alias{get_submissions}
\title{Get all submissions to all Kaggle competitions.}
\usage{
get_submissions(kaggle_db, team_competitions, with_predicted_place = TRUE)
}
\arguments{
\item{team_competitions}{Used to label competition ids in submissions.
If not provided, team competitions will be generated from the provided
Kaggle db.}

\item{with_predicted_place}{Set to true and/or provide leaderboards to
augment submissions with a PredictedPlace column based on final
standings.}
}
\description{
Simplifies the Submissions table of the Kaggle db.
}

