% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kaggle.R
\name{get_team_competitions}
\alias{get_team_competitions}
\title{Create a map of team ids to competition ids.}
\usage{
get_team_competitions(kaggle_db)
}
\description{
Create a map of team ids to competition ids.
}

