% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kaggle.R
\name{label_place_groups}
\alias{label_place_groups}
\title{Label different place groups, i.e., FirstPlaceTeam.}
\usage{
label_place_groups(frame)
}
\description{
Label different place groups, i.e., FirstPlaceTeam.
}

