% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kaggle.R
\name{label_submission_bins}
\alias{label_submission_bins}
\title{Cut total submissions into bins and label with the middle value.}
\usage{
label_submission_bins(leaderboards, submission_bin_width = 10)
}
\description{
Cut total submissions into bins and label with the middle value.
}

