% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kaggle.R
\name{label_submission_interval_bins}
\alias{label_submission_interval_bins}
\title{Cut submission intervals into bins and label with the middle time value.}
\usage{
label_submission_interval_bins(leaderboards, interval_bin_width)
}
\arguments{
\item{interval_bin_width}{Number of seconds to include in the bin. Defaults
to one day.}
}
\description{
Cut submission intervals into bins and label with the middle time value.
}

