% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/leaderboards.R
\name{make_leaderboards}
\alias{make_leaderboards}
\title{Create Kaggle leaderboards from submissions.}
\usage{
make_leaderboards(kaggle_db, submissions, team_sizes,
  with_submission_intervals = TRUE, with_competition_intervals = TRUE,
  with_relative_submissions = TRUE, label_team_types = TRUE)
}
\arguments{
\item{kaggle_db}{sqlite connector. Required unless submissions and team sizes
are both provided.}

\item{submissions}{data frame of Kaggle submissions. If submissions aren't
provided, they will be created.}

\item{team_sizes}{data frame map of Kaggle team ids to team sizes. Required
if connection to Kaggle db is not provided.}

\item{with_submission_intervals}{Logical. Should submission intervals
be calculated? Defaults to TRUE. Calls
\code{\link{calculate_submission_intervals}} to create the intervals.}

\item{with_competition_intervals}{Logical. Should start and end time
intervals be calculated for each competition/ Defaults to TRUE.
Calls \code{\link{get_competition_intervals}} to calculate the
competition intervals.}

\item{with_relative_submissions}{Logical. Should an additional column
be created calculating the number of submissions made for each team
relative to the team that took first place? Defaults to TRUE.
Calls \code{\link{calculate_relative_submissions}} to calculate
the relative submissions.}

\item{label_team_types}{Should submissions on leaderboards be assigned
team types in one of the four quadrants of the number submissions by
total time grid. See \code{\link{determine_team_types}}.}
}
\description{
Provide either submissions or a connection to the Kaggle db.
}
\examples{
kaggle_db <- connect_kaggle()
leaderboards <- make_leaderboards(kaggle_db)

}

