% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kaggle.R
\name{make_time_scale}
\alias{make_time_scale}
\title{Create a ggplot2 scale object for time variable in seconds.}
\usage{
make_time_scale(name, breaks_days, scale_obj = "scale_y_continuous", ...)
}
\description{
Create a ggplot2 scale object for time variable in seconds.
}

