% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/team-types.R
\name{recode_team_type}
\alias{recode_team_type}
\title{Recode team type for labels and contrast coding.}
\usage{
recode_team_type(frame)
}
\description{
If no data is provided, the team type map is returned.
}

