% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kaggle.R
\name{summarize_by_place}
\alias{summarize_by_place}
\title{Summarize team properties in each place.}
\usage{
summarize_by_place(leaderboards)
}
\description{
Each row is a place. For example, the row for Place == 1
describes the average number of TotalSubmissions and average
TeamSize for all teams that finished in this place.
}
\examples{
leaderboards \%>\% filter(Place <= 100) \%>\% summarize_by_place()

}

