% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kaggle.R
\name{summarize_teams_in_group}
\alias{summarize_teams_in_group}
\title{Summarize team performance in a group.}
\usage{
summarize_teams_in_group(grouped)
}
\description{
This functional sequence is meant to be passed a grouped data frame.
}
\examples{
leaderboards \%>\% group_by(TeamSize) \%>\% summarize_teams_in_group()

}

