% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kaggle.R
\name{make_leaderboards}
\alias{make_leaderboards}
\title{Create Kaggle leaderboards from submissions.}
\usage{
make_leaderboards(kaggle_db, submissions, team_sizes,
  with_submission_intervals = TRUE, with_competition_intervals = TRUE,
  with_relative_submissions = TRUE, label_team_types = TRUE)
}
\arguments{
\item{kaggle_db}{sqlite connector. Required unless submissions and team sizes
are both provided.}

\item{submissions}{data frame of Kaggle submissions. If submissions aren't
provided, they will be created.}

\item{team_sizes}{data frame map of Kaggle team ids to team sizes. Required
if connection to Kaggle db is not provided.}

\item{with_submission_intervals}{Logical. Should submission intervals
be calculated? Defaults to true. Calls
\code{\link{calculate_submission_intervals}} to create the intervals.}

\item{label_team_types}{Should submissions on leaderboards be assigned
team types in one of the four quadrants of the number submissions by
total time grid. See \code{\link{determine_team_types}}.}
}
\description{
Provide either submissions or a connection to the Kaggle db.
}
\examples{
submissions <- get_submissions(kaggle_db)
leaderboards <- make_leaderboards(submissions)
leaderboards <- make_leaderboards(kaggle_db = kaggle_db)

}

